+++ code/kdTree/parallelApproved/adaptBins_old.cpp
@@ -19,11 +19,11 @@ using namespace std;
 // Function: importFiles
 //
 
-void adaptBins_old( double *binE, int *binC, int numWorkers, int step ) {
+void adaptBins_old( float *binE, int *binC, int numWorkers, int step ) {
 	// binE = bin edges, binC = bin counts
 	
-	double *diff = new double[numWorkers+1];
-	double dC, dB, scale;
+	float *diff = new float[numWorkers+1];
+	float dC, dB, scale;
 	
 	for( int i = 1; i < numWorkers; i++ ) {
