+++ code/kdTree/parallelApproved/binData.h
@@ -10,7 +10,7 @@
 // Function: binDAta
 //
 
-void binData( double *data, double *binE, int myRank, int sortInd, 
+void binData( float *data, float *binE, int myRank, int sortInd, 
 	int numWorkers, int numPoints, int *binI, int *binC);
