+++ code/kdTree/parallelApproved/binData_noBlocks.cpp
@@ -37,7 +37,7 @@
 // returns binC, updates binI
 //
 
-void binData( double *data, double *binE, int myRank, int numWorkers, int numPoints, int *binI, int *binC) {
+void binData( float *data, float *binE, int myRank, int numWorkers, int numPoints, int *binI, int *binC) {
 	// binE = bin edges, binI = bin edge indices, binC = bin counts
 	
 	int test, halves, curr, last, ind;
