+++ code/kdTree/parallelApproved/max.h
@@ -5,7 +5,7 @@
 //
 
 // ****May need to make a limit for start and end point for the search for min max****
-double max(double *myArray, int rows, int cols, int maxThisColumn);
+float max(float *myArray, int rows, int cols, int maxThisColumn);
