+++ code/kdTree/parallelApproved/min.h
@@ -5,7 +5,7 @@
 //
 
 // ****May need to make a limit for start and end point for the search for min max****
-double min(double *myArray, int rows, int cols, int minThisColumn);
+float min(float *myArray, int rows, int cols, int minThisColumn);
