+++ code/kdTree/parallelApproved/receiveMinMax.cpp
@@ -29,19 +29,19 @@
 // Returns:
 //   Nothing
 
-void receiveMinMax(int sourceRank, double *min, double *max, MPI_Comm comm) {
+void receiveMinMax(int sourceRank, float *min, float *max, MPI_Comm comm) {
 
 	MPI_Status status;
-	auto minMax = new double[2];
+	auto minMax = new float[2];
 
 	minMax[_MIN_] = 0.0;
 	minMax[_MAX_] = 0.0;
 
-	int result = MPI_Recv(minMax, 2, MPI_DOUBLE, sourceRank,
+	int result = MPI_Recv(minMax, 2, MPI_FLOAT, sourceRank,
 		mpi_Tag_SendMinMax, comm, &status);
 
 	int counter;
-	MPI_Get_count(&status, MPI_DOUBLE, &counter);
+	MPI_Get_count(&status, MPI_FLOAT, &counter);
 
 	min[0] = minMax[_MIN_];
 	max[0] = minMax[_MAX_];
