+++ code/kdTree/parallelApproved/searchTree.h
@@ -20,6 +20,6 @@ using namespace std;
 // Function: buildTree
 //
 
-void searchTree( double *point, double rad, double *data[], int *rows, int cols, Tree *tree, MPI_Comm comm, int myRank, int numNodes, string name );
+void searchTree( float *point, float rad, float *data[], int *rows, int cols, Tree *tree, MPI_Comm comm, int myRank, int numNodes, string name );
