+++ code/kdTree/parallelApproved/sortData.cpp
@@ -44,26 +44,26 @@ int compareZ(const void *a, const void *b);
 // comparison methods (compareX, compareY, compareZ, compareIndex).
 //
 
-void sortData(double *data, int numCols, int numRows, int dimension)
+void sortData(float *data, int numCols, int numRows, int dimension)
 {
 	// auto start = std::chrono::system_clock::now();
 
 	switch (dimension)
 	{
 	case _X_:
-		qsort(data, numRows, numCols * sizeof(double), compareX);
+		qsort(data, numRows, numCols * sizeof(float), compareX);
 		break;
 
 	case _Y_:
-		qsort(data, numRows, numCols * sizeof(double), compareY);
+		qsort(data, numRows, numCols * sizeof(float), compareY);
 		break;
 
 	case _Z_:
-		qsort(data, numRows, numCols * sizeof(double), compareZ);
+		qsort(data, numRows, numCols * sizeof(float), compareZ);
 		break;
 
 	case _INDEX_:
-		qsort(data, numRows, numCols * sizeof(double), compareIndex);
+		qsort(data, numRows, numCols * sizeof(float), compareIndex);
 		break;
 
 	default:
@@ -72,7 +72,7 @@ void sortData(double *data, int numCols, int numRows, int dimension)
 	}
 
 	// auto end = std::chrono::system_clock::now();
-	// std::chrono::duration<double> elapsed = end - start;
+	// std::chrono::duration<float> elapsed = end - start;
 
 	// std::cout << "Sorted " << numRows << " rows in " << elapsed.count() << " seconds" << std::endl;
 
@@ -95,13 +95,13 @@ void sortData(double *data, int numCols, int numRows, int dimension)
 
 int compareX(const void *a, const void *b)
 {
-	if (static_cast<const double *>(a)[_X_] < static_cast<const double *>(b)[_X_])
+	if (static_cast<const float *>(a)[_X_] < static_cast<const float *>(b)[_X_])
 		return -1;
 
-	if (static_cast<const double *>(a)[_X_] == static_cast<const double *>(b)[_X_])
+	if (static_cast<const float *>(a)[_X_] == static_cast<const float *>(b)[_X_])
 		return 0;
 
-	if (static_cast<const double *>(a)[_X_] > static_cast<const double *>(b)[_X_])
+	if (static_cast<const float *>(a)[_X_] > static_cast<const float *>(b)[_X_])
 		return 1;
 
 	return 0;
@@ -125,13 +125,13 @@ int compareX(const void *a, const void *b)
 
 int compareIndex(const void *a, const void *b)
 {
-	if (static_cast<const double *>(a)[_INDEX_] < static_cast<const double *>(b)[_INDEX_])
+	if (static_cast<const float *>(a)[_INDEX_] < static_cast<const float *>(b)[_INDEX_])
 		return -1;
 
-	if (static_cast<const double *>(a)[_INDEX_] == static_cast<const double *>(b)[_INDEX_])
+	if (static_cast<const float *>(a)[_INDEX_] == static_cast<const float *>(b)[_INDEX_])
 		return 0;
 
-	if (static_cast<const double *>(a)[_INDEX_] > static_cast<const double *>(b)[_INDEX_])
+	if (static_cast<const float *>(a)[_INDEX_] > static_cast<const float *>(b)[_INDEX_])
 		return 1;
 	
 	return 0;
@@ -154,13 +154,13 @@ int compareIndex(const void *a, const void *b)
 
 int compareZ(const void *a, const void *b)
 {
-	if (static_cast<const double *>(a)[_Z_] < static_cast<const double *>(b)[_Z_])
+	if (static_cast<const float *>(a)[_Z_] < static_cast<const float *>(b)[_Z_])
 		return -1;
 
-	if (static_cast<const double *>(a)[_Z_] == static_cast<const double *>(b)[_Z_])
+	if (static_cast<const float *>(a)[_Z_] == static_cast<const float *>(b)[_Z_])
 		return 0;
 
-	if (static_cast<const double *>(a)[_Z_] > static_cast<const double *>(b)[_Z_])
+	if (static_cast<const float *>(a)[_Z_] > static_cast<const float *>(b)[_Z_])
 		return 1;
 	
 	return 0;
@@ -182,13 +182,13 @@ int compareZ(const void *a, const void *b)
 
 int compareY(const void *a, const void *b)
 {
-	if (static_cast<const double *>(a)[_Y_] < static_cast<const double *>(b)[_Y_])
+	if (static_cast<const float *>(a)[_Y_] < static_cast<const float *>(b)[_Y_])
 		return -1;
 
-	if (static_cast<const double *>(a)[_Y_] == static_cast<const double *>(b)[_Y_])
+	if (static_cast<const float *>(a)[_Y_] == static_cast<const float *>(b)[_Y_])
 		return 0;
 
-	if (static_cast<const double *>(a)[_Y_] > static_cast<const double *>(b)[_Y_])
+	if (static_cast<const float *>(a)[_Y_] > static_cast<const float *>(b)[_Y_])
 		return 1;
 	
 	return 0;
